Move parse_glyph_info to fontra.rs
In preparation for updating and integrating it into Font.
